+++ packages/expo-camera/ios/EXCamera/EXCameraCameraPermissionRequester.h
@@ -1,0 +1,8 @@
+// Copyright 2016-present 650 Industries. All rights reserved.
+
+#import <UMPermissionsInterface/UMPermissionsInterface.h>
+
+@interface EXCameraCameraPermissionRequester : NSObject<UMPermissionsRequester>
+
+@end
+