+++ packages/expo-camera/ios/EXCamera/EXCameraCameraPermissionRequester.m
@@ -1,13 +1,13 @@
 // Copyright 2016-present 650 Industries. All rights reserved.
 
-#import <EXBarCodeScanner/EXBarCodeCameraRequester.h>
+#import <EXCamera/EXCameraCameraPermissionRequester.h>
 #import <UMCore/UMDefines.h>
 #import <UMPermissionsInterface/UMPermissionsInterface.h>
 
 #import <AVFoundation/AVFoundation.h>
 
 
-@implementation EXBareCodeCameraRequester
+@implementation EXCameraCameraPermissionRequester
 
 + (NSString *)permissionType {
   return @"camera";
@@ -18,8 +18,9 @@
   AVAuthorizationStatus systemStatus;
   UMPermissionStatus status;
   NSString *cameraUsageDescription = [[NSBundle mainBundle] objectForInfoDictionaryKey:@"NSCameraUsageDescription"];
+  
   if (!cameraUsageDescription) {
-    UMFatal(UMErrorWithMessage(@"This app is missing 'NSCameraUsageDescription', so video services will fail. Add this entry to your bundle's Info.plist."));
+    UMFatal(UMErrorWithMessage(@"This app is missing NSCameraUsageDescription, so video services will fail. Add this entry to your bundle's Info.plist."));
     systemStatus = AVAuthorizationStatusDenied;
   } else {
     systemStatus = [AVCaptureDevice authorizationStatusForMediaType:AVMediaTypeVideo];
@@ -51,3 +52,4 @@
 }
 
 @end
+