+++ packages/expo-camera/ios/EXCamera/EXCameraMicrophonePermissionRequester.m
@@ -1,28 +1,29 @@
 // Copyright 2016-present 650 Industries. All rights reserved.
 
-#import <EXBarCodeScanner/EXBarCodeCameraRequester.h>
+#import <EXCamera/EXCameraMicrophonePermissionRequester.h>
 #import <UMCore/UMDefines.h>
 #import <UMPermissionsInterface/UMPermissionsInterface.h>
 
 #import <AVFoundation/AVFoundation.h>
 
 
-@implementation EXBareCodeCameraRequester
+@implementation EXCameraMicrophonePermissionRequester
 
 + (NSString *)permissionType {
-  return @"camera";
+  return @"microphone";
 }
 
 - (NSDictionary *)getPermissions
 {
   AVAuthorizationStatus systemStatus;
   UMPermissionStatus status;
-  NSString *cameraUsageDescription = [[NSBundle mainBundle] objectForInfoDictionaryKey:@"NSCameraUsageDescription"];
-  if (!cameraUsageDescription) {
-    UMFatal(UMErrorWithMessage(@"This app is missing 'NSCameraUsageDescription', so video services will fail. Add this entry to your bundle's Info.plist."));
+  NSString *microphoneUsageDescription = [[NSBundle mainBundle] objectForInfoDictionaryKey:@"NSMicrophoneUsageDescription"];
+    
+  if (!microphoneUsageDescription) {
+    UMFatal(UMErrorWithMessage(@"This app is missing NSMicrophoneUsageDescription, so audio services will fail. Add this entry to your bundle's Info.plist."));
     systemStatus = AVAuthorizationStatusDenied;
   } else {
-    systemStatus = [AVCaptureDevice authorizationStatusForMediaType:AVMediaTypeVideo];
+    systemStatus = [AVCaptureDevice authorizationStatusForMediaType:AVMediaTypeAudio];
   }
   switch (systemStatus) {
     case AVAuthorizationStatusAuthorized:
@@ -44,10 +45,12 @@
 - (void)requestPermissionsWithResolver:(UMPromiseResolveBlock)resolve rejecter:(UMPromiseRejectBlock)reject
 {
   UM_WEAKIFY(self)
-  [AVCaptureDevice requestAccessForMediaType:AVMediaTypeVideo completionHandler:^(BOOL granted) {
+  [AVCaptureDevice requestAccessForMediaType:AVMediaTypeAudio completionHandler:^(BOOL granted) {
     UM_STRONGIFY(self)
     resolve([self getPermissions]);
   }];
 }
 
 @end
+
+